requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
yfinance
pytz
requests
google-generativeai
